Fuel Collection Game
A simple 2D array-based game where the player controls a ship that moves left and right to collect fuel. The objective is to collect all the fuel without running out of fuel. If the ship collects all the fuel in time, the game is won. If the ship runs out of fuel before collecting all the fuel, the game is lost.