Linked List as an ADT
The LinkedListADT project implements a linked list data structure as an Abstract Data Type (ADT) in C++. The project is designed using separate compilation, with the linked list class interface defined in a header file and its implementation in a source file. The linked list will support the following operations:

Inserting at the back: Adds a new element to the end of the list.
Inserting at the front: Adds a new element to the beginning of the list.
Searching: Finds an element in the list.
Deleting: Removes an element from the list.
Printing the list in reverse: Displays the list elements in reverse order. The project will be developed and tested using Codeblocks as the Integrated Development Environment (IDE). The implementation will follow a modular approach, allowing for easy modifications and reusability.